refactor(hookruntime): address agent-review findings on the rename
- move the wire-format contract prose from a file header (which godoc
  silently adopted as a malformed package doc) onto DecodeStandardEvent,
  and document EncodeStandardResult's pre-tool-use-only enforcement
- rename stale Claude-format test names in primeagent tests
- align the Cowork coverage row with the standard-format terminology
- move the two direct-codec TestEncodeStandardResult tests from the
  claude adapter package into hookruntime/standard_test.go
- codex decode errors now use the agentName parameter like the standard
  codec instead of a hardcoded prefix
