+++ docs/coverage.md
@@ -13,7 +13,7 @@ supported hook. What varies is the event surface the agent exposes.
 | Agent | Events recorded | Can block | Installation | Status and limits |
 | --- | --- | --- | --- | --- |
 | Claude Code | Session start/end; pre-tool-use; successful and failed post-tool-use | Pre-tool-use only | `kontext setup` installs a managed Claude Code hook configuration | Supported. Session lifecycle hooks run asynchronously, so they record context but do not hold up the agent. Claude's configured hook surface does not include prompt-submission events. |
-| Claude Cowork | The Claude Code hook event format: session start/end, pre-tool-use, and successful or failed post-tool-use | Pre-tool-use only | Configure the Cowork hook to invoke `kontext hook --agent cowork …` | Supported at the hook-protocol level. Cowork runs Claude Code inside a per-session VM; deployment and hook configuration are owned by that environment, not by `kontext setup`. |
+| Claude Cowork | The standard hook wire format (Claude Code’s hook event format): session start/end, pre-tool-use, and successful or failed post-tool-use | Pre-tool-use only | Configure the Cowork hook to invoke `kontext hook --agent cowork …` | Supported at the hook-protocol level. Cowork runs Claude Code inside a per-session VM; deployment and hook configuration are owned by that environment, not by `kontext setup`. |
 | Codex | Session start; pre-tool-use; post-tool-use; user-prompt-submit; stop | Pre-tool-use. Codex can also receive a block result for post-tool-use or user-prompt-submit events. | `kontext setup` adds user hooks and enables the Codex hooks feature; the hooks must then be trusted in Codex. | Supported. User-scoped setup is currently macOS-only. Hook status verifies the configuration, not that a hook has executed. |
 | Prime Agent | Session start/end; pre-tool-use; successful and failed post-tool-use; user-prompt-submit; stop | Pre-tool-use only | `kontext setup` installs a managed extension into `~/.prime/agent/extensions` when Prime Agent is detected | Supported via Prime Agent's extension system, which exposes a synchronous `tool_call` callback. The managed extension forwards events in the standard hook wire format (originally the Claude Code hook-input schema) to `kontext hook --agent prime-agent` and fails open (with a visible warning) when the local runtime is unavailable. Prompt-submit events are recorded but not blocked in this integration. |
 | Other agents | — | — | — | Not yet a shipped integration. A compatible hook adapter can use the local runtime, but it is not covered by this matrix until its event contract and enforcement behavior are documented and tested. |
