+++ internal/agent/claude/claude_test.go
@@ -6,7 +6,6 @@ import (
 	"testing"
 
 	"github.com/kontext-security/kontext-cli/internal/hook"
-	"github.com/kontext-security/kontext-cli/internal/hookruntime"
 )
 
 func TestDecodeHookInputPreservesOptionalMetadata(t *testing.T) {
@@ -203,42 +202,6 @@ func TestEncodeAllowIncludesUpdatedInput(t *testing.T) {
 	}
 	if !strings.Contains(string(out), "suppressOutput") {
 		t.Fatalf("EncodeHookResult() = %s, want suppressOutput", out)
-	}
-}
-
-func TestEncodeStandardResultMapsUnsupportedDecisionToDeny(t *testing.T) {
-	t.Parallel()
-
-	out, err := hookruntime.EncodeStandardResult("PreToolUse", hook.Result{
-		Decision: hook.Decision("ask"),
-		Reason:   "approval required",
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	if !strings.Contains(string(out), `"permissionDecision":"deny"`) {
-		t.Fatalf("output = %s", string(out))
-	}
-	if !strings.Contains(string(out), `"permissionDecisionReason":"approval required"`) {
-		t.Fatalf("output = %s", string(out))
-	}
-}
-
-func TestEncodeStandardResultOmitsDecisionForPostToolUse(t *testing.T) {
-	t.Parallel()
-
-	out, err := hookruntime.EncodeStandardResult("PostToolUse", hook.Result{
-		Decision: hook.DecisionDeny,
-		Reason:   "telemetry only",
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	if strings.Contains(string(out), "permissionDecision") {
-		t.Fatalf("output = %s, want no PreToolUse permission decision", string(out))
-	}
-	if !strings.Contains(string(out), `"suppressOutput":true`) {
-		t.Fatalf("output = %s, want suppressed output", string(out))
 	}
 }
